backend/pyproject.toml
[tool.black]
line-length = 99
target-version = ['py37']
include = '\.pyi?$'
exclude = '''
(
  /(
    | \.mypy_cache
    | \.venv
  )/
  | migrations
)
'''

[tool.poetry]
name = "disfactory"
version = "0.0.0"
description = ""
authors = []

[tool.poetry.dependencies]
python = "^3.6"
django = "2.2.13"
psycopg2-binary = "2.8.3"
djangorestframework = "3.9.2"
python-dotenv = "^0.14.0"
requests = "^2.24.0"
pillow = "^8.1.1"
django-cors-headers = "^3.5.0"
gunicorn = "^20.0.4"
django-admin-rangefilter = "^0.6.2"
django-map-widgets = "^0.3.0"
drf-yasg = "^1.17.1"
python-docx = "^0.8.10"
docxcompose = "^1.2.0"
django-q = "1.0.2"
django-import-export = "^2.4.0"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
freezegun = "^1.0.0"
pylint = "^2.6.0"
pylint-django = "^2.3.0"
flake8 = "^3.8.3"
pytest-django = "^4.1.0"
